vast/requirements_vast.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.37.2
accelerate==0.27.0
bitsandbytes==0.42.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
requests==2.31.0
scipy
networkx
tiktoken
protobuf
sentencepiece
pillow
qwen-vl-utils